camtraptor: Explore and Visualize Camera Trap Data Packages

Explore and visualize Camera Trap Data Packages ('Camtrap DP', https://camtrap-dp.tdwg.org/). camtraptor offers a step-by-step workflow to read Camtrap DP files, filter data of interest, summarize information (e.g. number of observed species) and visualize this per deployment on an interactive map. You can also use it to transform data for analysis in 'camtrapR'.

See also

Author

Maintainer: Damiano Oldoni damiano.oldoni@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO))

Authors:

Damiano Oldoni damiano.oldoni@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO))

Peter Desmet peter.desmet@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO))

Sanne Govaert sanne.govaert@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO))

Pieter Huybrechts pieter.huybrechts@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO))

Wolf Missotten wolfmissotten@gmail.com (ORCID) (affiliation: Ghent University)

Other contributors:

Martijn Bollen martijn.bollen@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]

Jim Casaer jim.casaer@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]

Nele Mullens (ORCID) (affiliation: Regionaal Landschap Dijleland vzw) [contributor]

Raphaël Nussbaumer rafnuss@gmail.com (ORCID) (affilication: Swiss Ornithological Institute) [contributor]

Thierry Onkelinx thierry.onkelinx@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]

Lynn Pallemaerts lynn.pallemaerts@inbo.be (ORCID) (affiliation: Research Institute for Nature and Forest (INBO)) [contributor]

Research Institute for Nature and Forest (INBO) (ROR) [copyright holder]

Research Foundation - Flanders (https://lifewatch.be) [funder]