Get number of identified species for each deployment

This function is deprecated. Use summarize_observations() instead.

Usage

get_n_species(x, ...)

Arguments

x

Camera trap data package object, as returned by camtrapdp::read_camtrapdp().

...

filter predicates for filtering on deployments are not supported anymore and an error is returned. Anything else throws a deprecation warning and will be ignored. Please use filter_deployments() to filter on deployments.

Value

A tibble data frame with the following columns:

deploymentID: Deployment unique identifier.

n: Number of observed and identified species.

See also

Examples

x <- example_dataset()
# Get number of species
get_n_species(x)
#> Warning: `get_n_species()` was deprecated in camtraptor 1.0.0.
#> ℹ Please use `summarize_observations(x, group_by = c("deploymentID",
#>   "latitude"), "longitude"), "scientificName")` instead.
#> ℹ The deprecated feature was likely used in the camtraptor package.
#>   Please report the issue at <https://github.com/inbo/camtraptor/issues>.
#> # A tibble: 4 × 2
#> # Groups:   deploymentID [4]
#>   deploymentID     n
#>   <chr>        <int>
#> 1 00a2c20d         3
#> 2 29b7d356         3
#> 3 577b543a         3
#> 4 62c200a9         2