Get species

Gets all identified species.

Usage

get_species(package = NULL, datapkg = lifecycle::deprecated())

Arguments

package

Camera trap data package object, as returned by read_camtrap_dp().

datapkg

Deprecated. Use package instead.

Value

A tibble data frame with all scientific names and vernacular names of the identified species.

See also

Examples

get_species(mica)
#> # A tibble: 9 × 5
#>   taxonID taxonIDReference                scientificName     vernacula…¹ verna…²
#>   <chr>   <chr>                           <chr>              <chr>       <chr>  
#> 1 DGP6    https://www.catalogueoflife.org Anas platyrhynchos mallard     wilde …
#> 2 DGPL    https://www.catalogueoflife.org Anas strepera      gadwall     krakee…
#> 3 32FH    https://www.catalogueoflife.org Ardea              great hero… reigers
#> 4 GCHS    https://www.catalogueoflife.org Ardea cinerea      grey heron  blauwe…
#> 5 RQPW    https://www.catalogueoflife.org Castor fiber       Eurasian b… bever  
#> 6 6MB3T   https://www.catalogueoflife.org Homo sapiens       human       mens   
#> 7 3Y9VW   https://www.catalogueoflife.org Martes foina       beech mart… steenm…
#> 8 44QYC   https://www.catalogueoflife.org Mustela putorius   European p… bunzing
#> 9 5BSG3   https://www.catalogueoflife.org Vulpes vulpes      red fox     vos    
#> # … with abbreviated variable names ¹​vernacularNames.en, ²​vernacularNames.nl