Transform efforts to common units.

This function is useful for visualization and communication purposes. Efforts are duration objects and so they are always expressed in seconds. Although they are also returned on screen in common units, e.g. "2594308s (~4.29 weeks)", the values in seconds are used in color scales, not very handy to be interpreted. Converting them in the most suitable time unit is also useful for communication purposes (reports, research articles, ...). Obviously the conversion can be not always precise, e.g. a month is not always 30 days long.

Usage

transform_effort_to_common_units(effort, unit)

Arguments

effort

a vector of duration objects

unit

common unit to express duration objects. One of:

second

minute

hour

day

month

year

Value

a numeric vector

Examples

# create efforts (durations) to transform
efforts <- c(lubridate::duration("2hours 2minutes 1second"),
            lubridate::duration("3days 2hours"))
# transform effort to hours
transform_effort_to_common_units(efforts, "hour")
#> [1]  2.033611 74.000000
# transform effort to days
transform_effort_to_common_units(efforts, "day")
#> [1] 0.0847338 3.0833333