Transform camera trap data to Darwin Core

Transforms a published Camera Trap Data Package (Camtrap DP) to Darwin Core CSV and EML files that can be uploaded to a GBIF IPT for publication. A meta.xml file is not created.

write_dwc(
  package,
  directory = ".",
  doi = package$id,
  contact = NULL,
  rights_holder = package$rightsHolder
)

Arguments

package

A Camtrap DP, as read by read_camtrap_dp().

directory

Path to local directory to write files to.

doi

DOI of the original dataset, used to get metadata.

contact

Person to be set as resource contact and metadata provider. To be provided as a person().

rights_holder

Acronym of the organization owning or managing the rights over the data.

Value

CSV (data) and EML (metadata) files written to disk.

Metadata

Metadata are derived from the original dataset by looking up its doi in DataCite (example) and transforming these to EML. Uses movepub::datacite_to_eml() under the hood. The following properties are set:

title: Original title + [animal observations].

description: Automatically created first paragraph describing this is a derived dataset, followed by the original dataset description.

license: License of the original dataset.

creators: Creators of the original dataset.

contact: contact or first creator of the original dataset.

metadata provider: contact or first creator of the original dataset.

keywords: Keywords of the original dataset.

associated parties: Organizations as defined in package$organizations.

geographic coverage: Bounding box as defined package$spatial.

taxonomic coverage: Species as defined in package$taxonomic.

temporal coverage: Date range as defined in package$temporal.

project data: Title, identifier, description, and sampling design information as defined in package$project.

alternative identifier: DOI of the original dataset. This way, no new DOI will be created when publishing to GBIF.

external link: URL of the project as defined in package$project$path.

To be set manually in the GBIF IPT: type, subtype, update frequency, and publishing organization.

Not set: sampling methods and citations. Not applicable: collection data.

Data

package is expected to contain the resources deployments, media and observations. Their CSV data are loaded in to a SQLite database, transformed to Darwin Core using SQL and written to disk as CSV file(s).

Key features of the Darwin Core transformation:

TODO

Examples

# TODO